memqueue cleanup
Move the buffer and message handling from memqueue broker type
to the actual go routines handling the messaging.
