+++ libbeat/publisher/queue/memqueue/consume.go
@@ -24,10 +24,11 @@ type consumerStats struct {
 }
 
 type batch struct {
-	consumer *consumer
-	events   []publisher.Event
-	ack      *ackChan
-	state    ackState
+	consumer     *consumer
+	events       []publisher.Event
+	clientStates []clientState
+	ack          *ackChan
+	state        ackState
 }
 
 type ackState uint8
@@ -112,5 +113,5 @@ func (b *batch) ACK() {
 }
 
 func (b *batch) report() {
-	b.ack.ch <- batchAckRequest{}
+	b.ack.ch <- batchAckMsg{}
 }
